backend/requirements.txt
alembic==1.16.5
pytest-asyncio==0.26.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
asyncpg==0.31.0
beautifulsoup4==4.14.3
certifi==2026.2.25
click==8.1.8
exceptiongroup==1.3.1
fastapi==0.128.8
greenlet==3.2.5
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
python-dotenv==1.2.1
PyYAML==6.0.3
soupsieve==2.8.3
SQLAlchemy==2.0.47
starlette==0.49.3
tomli==2.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.39.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
